+++ src/components/main/main.module.css
@@ -1,6 +1,7 @@
 .main{
     display: flex;
-    justify-content: space-around;  
+    justify-content: center;  
+    gap: 40px;
 }
 
 .burgerIngridients{
