+++ src/query.txt
@@ -1,11 +1,6 @@
-как задать стили для компонента!!!
-поместить все в одно окно
-нормальная прокрутка
-пробелы у кристалов
-лого по середине
-чтоб не разъезжалось при разжирнении
+
 добавление ингредиентов в лист
 присобачить таб
-Соус Spicy-X разжирнить
-пин для ингредиентов
 проверка типов
+скроллбар для разных размеров экрана
+ логика подсчета суммы
